% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_imu.R
\name{plot_gyr}
\alias{plot_gyr}
\title{Plot gyroscope}
\usage{
plot_gyr(data, max_points = 10000)
}
\arguments{
\item{data}{A data.frame with columns:
* datetime_local (POSIXct)
* gyrx (double)
* gyry (double)
* gyrz (double)}
}
\value{
A ggplot object
}
\description{
\code{plot_gyr} plots x, y, and z-axis gyroscope
}
