% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_imu.R
\name{plot_mag}
\alias{plot_mag}
\title{Plot magnetometer}
\usage{
plot_mag(data, max_points = 10000)
}
\arguments{
\item{data}{A data.frame with columns:
* datetime_local (POSIXct)
* magx (double)
* magy (double)
* magz (double)}
}
\value{
A ggplot object
}
\description{
\code{plot_mag} plots x, y, and z-axis magnetometer
}
